+++ src/extension/AL Code Outline Ext/fullSyntaxTreeNodeKind.ts
@@ -35,6 +35,9 @@ export class FullSyntaxTreeNodeKind {
     public static getXmlPortObject(): string {
         return 'XmlPortObject';
     }
+    public static getQueryObject(): string {
+        return 'QueryObject';
+    }
     public static getEnumType(): string {
         return 'EnumType';
     }
@@ -127,6 +130,12 @@ export class FullSyntaxTreeNodeKind {
     }
     public static getCaseStatement(): string {
         return 'CaseStatement';
+    }
+    public static getCaseLine(): string {
+        return 'CaseLine';
+    }
+    public static getCaseElse(): string {
+        return 'CaseElse';
     }
     public static getWithStatement(): string {
         return 'WithStatement';
@@ -223,6 +232,9 @@ export class FullSyntaxTreeNodeKind {
     }
     public static getObjectId(): string {
         return 'ObjectId';
+    }
+    public static getObjectReference(): string {
+        return 'ObjectReference';
     }
     public static getAllStatementKinds(): string[] {
         return [