+++ src/extension/Create Procedure/CodeActionCreatorFactory.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { DiagnosticAnalzyer } from '../Utils/diagnosticAnalyzer';
+import { DiagnosticAnalyzer } from '../Utils/diagnosticAnalyzer';
 import { SupportedDiagnosticCodes } from './supportedDiagnosticCodes';
 import { ICodeActionCreator } from './Code Action Creator/ICodeActionCreator';
 import { CodeActionCreatorAL0118 } from './Code Action Creator/CodeActionCreatorAL0118';
@@ -8,7 +8,7 @@ import { CodeActionCreatorAL0499 } from './Code Action Creator/CodeActionCreator
 
 export class CodeActionCreatorFactory {
     public static getInstances(document: vscode.TextDocument, range: vscode.Range): ICodeActionCreator[] {
-        let diagnostics: vscode.Diagnostic[] = new DiagnosticAnalzyer().getValidDiagnosticOfCurrentPositionToCreateProcedure(document, range);
+        let diagnostics: vscode.Diagnostic[] = new DiagnosticAnalyzer().getValidDiagnosticOfCurrentPositionToCreateProcedure(document, range);
         if (diagnostics.length === 0) {
             return [];
         }